+++ views/addreview.html
@@ -12,7 +12,7 @@
 			<% if (addResult == "success") { %>
 			<div class="alert alert-success" role="alert">Your review was successfully added. Thanks!</div>
 			<% } %>
-			<form action="/add" method="POST" class="mb-4">
+			<form method="POST" class="mb-4">
 				<div class="form-group">
 					<label for="module">Module</label>
 					<select
