+++ views/editreview.html
@@ -16,7 +16,7 @@
 
 			<!-- Form -->
 			<!-- overrides method with query string -->
-			<form action="/review/<%= review.id %>/delete?_method=DELETE" method="POST" class="mb-4">
+			<form action="/reviews/<%= review.id %>/delete?_method=DELETE" method="POST" class="mb-4">
 				<input
 					class="btn btn-danger"
 					type="submit"
@@ -25,7 +25,7 @@
 				/>
 			</form>
 			<!-- overrides method with query string -->
-			<form action="/review/<%- review.id %>/update?_method=PUT" method="POST" class="mb-4">
+			<form action="/reviews/<%- review.id %>/update?_method=PUT" method="POST" class="mb-4">
 				<div class="form-group">
 					<label for="module">Module</label>
 					<select
