+++ views/index.html
@@ -17,7 +17,7 @@
 			<!-- Page content begins below this -->
 			<h1><%- heading %></h1>
 
-			<a href="/add" class="btn btn-primary">Add a review</a>
+			<a href="/reviews/add" class="btn btn-primary">Add a review</a>
 			<table data-toggle="table" data-search="true" class="pb-10">
 				<thead>
 					<tr>
