+++ views/reviews.html
@@ -8,7 +8,7 @@
 		<div class="container">
 			<!-- Page content begins below this -->
 			<h1><%- heading %></h1>
-			<a href="/add<%- typeof filteredModule !== 'undefined' ? '?course_id=' + filteredModule : '' %>"
+			<a href="/reviews/add<%- typeof filteredModule !== 'undefined' ? '?course_id=' + filteredModule : '' %>"
 				class="btn btn-primary mb-4" >Add a review</a>
 			<% if (typeof reviews === 'undefined' || reviews.length < 1) { %>
 			<p>No reviews found for selected course or id.</p>
@@ -17,7 +17,7 @@
 			<div class="card mb-3 review-container">
 				<div class="card-body">
 					<h5 class="card-title">
-						<a href="/review/<%- review.id %>"> <%- review.course_id %>: <%- review.title %> </a>
+						<a href="/reviews/<%- review.id %>"> <%- review.course_id %>: <%- review.title %> </a>
 					</h5>
 					<h6 class="card-subtitle mb-2 text-muted"><%- review.timestamp %></h6>
 					<h6 class="card-subtitle mb-2 text-muted"><%- review.author %></h6>
@@ -27,7 +27,7 @@
 					<h6 class="card-subtitle mb-2">Workload: <%- review.workload %> hours / week</h6>
 					<h6 class="card-subtitle mb-2">Rating: <%- review.rating %> / 5</h6>
 					<% if (typeof user !== 'undefined' && user.id == review.user_id) { %>
-					<a class="btn btn-outline-secondary btn-sm" href="/review/<%- review.id %>/update">Edit</a>
+					<a class="btn btn-outline-secondary btn-sm" href="/reviews/<%- review.id %>/update">Edit</a>
 					<% } %>
 				</div>
 			</div>
